src/requirement.txt
cffi==1.6.0
cryptography==1.3.2
enum34==1.1.6
gitdb==0.6.4
GitPython==2.0.2
idna==2.1
ipaddress==1.0.16
paramiko==2.0.0
pyasn1==0.1.9
pycparser==2.14
scp==0.10.2
six==1.10.0
smmap==0.9.0
